Module Threadext.Delay

type t

val make : unit -> t

Blocks the calling thread for a given period of time with the option of returning early if someone calls 'signal'. Returns true if the full time period elapsed and false if signalled. Note that multple 'signals' are coalesced; 'signals' sent before 'wait' is called are not lost.

val wait : t -> float -> bool

Sends a signal to a waiting thread. See 'wait'

val signal : t -> unit